Revert "Update test runner for osx"
This reverts commit b02c91f594db14f5e7cda807f53118ef3750dc89.
