+++ tests/_utils/system.py
@@ -54,9 +54,7 @@ class System:
 
     """
 
-    def __init__(self, manager: multiprocessing.managers.SyncManager, *, dry_run: bool = False) -> None:
-        self.manager = manager
-        self.pool = multiprocessing.Pool(20)
+    def __init__(self, *, dry_run: bool = False) -> None:
         self.dry_run: bool = dry_run
 
     def run(
