+++ tests/_utils/test_stages/gpu.py
@@ -14,15 +14,17 @@
 #
 from __future__ import annotations
 
-from ... import FeatureType
-from ...config import Config
-from ...system import System
-from ...types import ArgList, EnvDict
-from ..test_stage import Shard, StageSpec, TestStage, adjust_workers
+from .. import FeatureType
+from ..config import Config
+from ..system import System
+from ..types import ArgList, EnvDict
+from .test_stage import Shard, StageSpec, TestStage, adjust_workers
+
+BLOAT_FACTOR = 1.5  # hard coded for now
 
 
-class OMP(TestStage):
-    """A test stage for exercising OpenMP features.
+class GPU(TestStage):
+    """A test stage for exercising GPU features.
 
     Parameters
     ----------
@@ -34,7 +36,7 @@ class OMP(TestStage):
 
     """
 
-    kind: FeatureType = "openmp"
+    kind: FeatureType = "cuda"
 
     args = ["-cunumeric:test"]
 
@@ -45,24 +47,29 @@ class OMP(TestStage):
 
     def shard_args(self, shard: Shard, config: Config) -> ArgList:
         return [
-            "--omps",
-            str(config.omps),
-            "--ompthreads",
-            str(config.ompthreads),
-            "--cpu-bind",
+            "--fbmem",
+            str(config.fbmem // (1024 * 1024)),  # accepts size in MB
+            "--gpus",
+            str(len(shard)),
+            "--gpu-bind",
             ",".join(str(x) for x in shard),
         ]
 
     def compute_spec(self, config: Config, system: System) -> StageSpec:
-        N = len(system.cpus)
-        omps, threads = config.omps, config.ompthreads
-        degree = N // (omps * threads + config.utility)
+        N = len(system.gpus)
+        degree = N // config.gpus
 
-        workers = adjust_workers(degree, config.requested_workers)
+        fbsize = min(gpu.total for gpu in system.gpus)
+        oversub_factor = int(fbsize // (config.fbmem * BLOAT_FACTOR))
+        workers = adjust_workers(
+            degree * oversub_factor, config.requested_workers
+        )
 
         # https://docs.python.org/3/library/itertools.html#itertools-recipes
         # grouper('ABCDEF', 3) --> ABC DEF
-        args = [iter(range(workers * omps * threads))] * (omps * threads)
-        shards = list(zip(*args))
+        args = [iter(range(degree * config.gpus))] * config.gpus
+        per_worker_shards = list(zip(*args))
+
+        shards = per_worker_shards * workers
 
         return StageSpec(workers, shards)
